guidelines format; moved newton_manager.py
