+++ source/isaaclab/isaaclab/envs/manager_based_rl_env.py
@@ -21,11 +21,14 @@ from isaaclab.utils.version import get_isaac_sim_version
 
 from .common import VecEnvStepReturn
 
+
 def _get_isaac_sim_version_str() -> str:
     try:
         return str(get_isaac_sim_version())
     except Exception:
         return "unknown"
+
+
 from .manager_based_env import ManagerBasedEnv
 from .manager_based_rl_env_cfg import ManagerBasedRLEnvCfg
 
@@ -246,7 +249,10 @@ class ManagerBasedRLEnv(ManagerBasedEnv, gym.Env):
         # -- compute observations (includes camera/tiled camera rendering)
         # note: done after reset to get the correct observations for reset envs
         if self.common_step_counter <= 3 or self.common_step_counter % 50 == 0:
-            print(f"[PERF][manager_based_rl_env] Computing observations (camera/tiled camera render) step #{self.common_step_counter}...", flush=True)
+            print(
+                f"[PERF][manager_based_rl_env] Computing observations (camera/tiled camera render) step #{self.common_step_counter}...",
+                flush=True,
+            )
         with Timer(name="render", msg="Rendering step took"):
             self.obs_buf = self.observation_manager.compute(update_history=True)
