+++ source/isaaclab/isaaclab/managers/__init__.py
@@ -15,6 +15,7 @@ from .command_manager import CommandManager, CommandTerm
 from .curriculum_manager import CurriculumManager
 from .event_manager import EventManager
 from .manager_base import ManagerBase, ManagerTermBase
+from .newton_manager import NewtonManager
 from .manager_term_cfg import (
     ActionTermCfg,
     CommandTermCfg,
